+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.20.0
+
+- Return `emitter` with events: `process` and `error`.
+- Stats are no longer logged.
+- Use `BIGINT` column type for `number` when creating table schema.
+
 # 0.19.0
 
 - Removed `clearCompletedOn`.